Conflicts:
	Source/src/org/coury/jfilehelpers/masterdetail/MasterDetailEngine.java

+++ Tests/src/org/coury/jfilehelpers/tests/converters/testobjects/DecimalType.java
@@ -1,5 +1,7 @@
 package org.coury.jfilehelpers.tests.converters.testobjects;
 
+import java.math.BigDecimal;
+
 import org.coury.jfilehelpers.annotations.DelimitedRecord;
 
 @DelimitedRecord("|")
@@ -7,6 +9,5 @@ public class DecimalType {
 	public int intField;
 	public float floatField;
 	public double doubleField;
-	// added just for C# compatibility
-	public float decimalField;
+	public BigDecimal decimalField;
 }
